Remove useless cast

+++ decoder/LAVVideo/decoders/msdk_mvc.cpp
@@ -499,7 +499,7 @@ HRESULT CDecMSDKMVC::ParseOffsetMetadata(const BYTE *buffer, int size, mfxU64 ti
   for (int o = 0; o < nOffsets; o++) {
     for (int f = 0; f < nFrames; f++) {
       if (o == 0) {
-        MediaSideData3DOffset offset = { (BYTE)nOffsets };
+        MediaSideData3DOffset offset = { nOffsets };
         GOP.offsets.push_back(offset);
       }
